perform refactoring
