+++ Booksies/app/src/main/java/com/example/booksies/controller/HomeFragment.java
@@ -1,30 +1,22 @@
 package com.example.booksies.controller;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
-import android.widget.EditText;
-import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.SearchView;
 import android.widget.Spinner;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.DefaultItemAnimator;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.booksies.model.Books;
 import com.example.booksies.model.FirestoreHandler;
-import com.example.booksies.model.MyAdapter;
 import com.example.booksies.R;
 import com.google.firebase.firestore.FirebaseFirestore;
