+++ Booksies/app/src/main/java/com/example/booksies/model/Books.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
  * Class models a book object.
  * This class keeps track of the attributes of a book object
  */
-public class Books {
+public class Books implements Identifiable {
     private String title;
     private String author;
     private String ISBN;
