--- Booksies/app/src/main/java/com/example/booksies/model/BooksActivity.java
@@ -1,6 +1,0 @@
-package com.example.booksies.model;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class BooksActivity extends AppCompatActivity {
-}
